Great Progress Bugfixing
All existing slash commands (except disband)
are all working correctly and cool and good.
Next step: fix disband_faction
